requirements.txt
streamlit
groq
elasticsearch
python-dotenv
yfinance
